UPD Double Alice

+++ main.cpp
@@ -7,7 +7,11 @@ int main()
 {
    cout << "Enter A and B: ";
 
-int a, b,max,min;
+
+
+
+double a, b,max,min;
+
